+++ voucher-project/src/main/java/com/example/voucherproject/common/config/AppConfig.java
@@ -1,14 +1,14 @@
 package com.example.voucherproject.common.config;
 
 import com.example.voucherproject.AllService;
-import com.example.voucherproject.common.file.CSVReader;
-import com.example.voucherproject.common.file.CSVWriter;
-import com.example.voucherproject.common.file.MyReader;
-import com.example.voucherproject.common.file.MyWriter;
-import com.example.voucherproject.common.io.ConsoleIn;
-import com.example.voucherproject.common.io.ConsoleOut;
-import com.example.voucherproject.common.io.Input;
-import com.example.voucherproject.common.io.Output;
+import com.example.voucherproject.common.io.console.ConsoleIn;
+import com.example.voucherproject.common.io.console.ConsoleOut;
+import com.example.voucherproject.common.io.console.Input;
+import com.example.voucherproject.common.io.console.Output;
+import com.example.voucherproject.common.io.file.CSVReader;
+import com.example.voucherproject.common.io.file.CSVWriter;
+import com.example.voucherproject.common.io.file.MyReader;
+import com.example.voucherproject.common.io.file.MyWriter;
 import com.example.voucherproject.user.service.UserService;
 import com.example.voucherproject.voucher.service.VoucherService;
 import org.springframework.context.annotation.Bean;
